refactor: 조회 rdb -> res
- scheduler를 이용해 1분 마다 조회수 동기화
- BoardHit 테이블 사용하지 않고 Board 테이블에 hit 컬럼 추가
- redis의 key-value 구조를 24시간 동안 키 값 유지로 조회수 중복 증가 방지

+++ src/main/java/chathealth/chathealth/dto/response/BoardResponse.java
@@ -28,7 +28,7 @@ public class BoardResponse {
 
     //조회수, 댓글수
     private int commentCount;
-    private int hit;
+    private long hit;
 
     private Boolean isWriter;
 }
